requirements.txt
boto3
Flask
Pillow
flask-cors